move a few hinting-related checks out of the googlefonts directory
* hinting_impact
* smart_dropout
* vttclean
* integer_ppem_if_hinted
