+++ tests/test_checks_googlefonts.py
@@ -1040,21 +1040,6 @@ def test_check_name_description_max_length():
     )
 
 
-def test_check_hinting_impact():
-    """Show hinting filesize impact."""
-    check = CheckTester("hinting_impact")
-
-    font = TEST_FILE("mada/Mada-Regular.ttf")
-    assert_results_contain(
-        check(font), INFO, "size-impact", "this check always emits an INFO result..."
-    )
-
-    font = TEST_FILE("rokkitt/Rokkitt-Bold.otf")
-    assert_results_contain(
-        check(font), INFO, "size-impact", "this check always emits an INFO result..."
-    )
-
-
 def test_check_name_version_format():
     """Version format is correct in 'name' table ?"""
     check = CheckTester("googlefonts:name/version_format")
@@ -2768,35 +2753,6 @@ def test_check_varfont_has_HVAR():
     assert_results_contain(check(ttFont), FAIL, "lacks-HVAR")
 
 
-def test_check_smart_dropout():
-    """Font enables smart dropout control in "prep" table instructions?"""
-    check = CheckTester("smart_dropout")
-
-    ttFont = TTFont(TEST_FILE("nunito/Nunito-Regular.ttf"))
-
-    # "Program at 'prep' table contains
-    #  instructions enabling smart dropout control."
-    assert_PASS(check(ttFont))
-
-    # "Font does not contain TrueType instructions enabling
-    #  smart dropout control in the 'prep' table program."
-    import array
-
-    ttFont["prep"].program.bytecode = array.array("B", [0])
-    assert_results_contain(check(ttFont), FAIL, "lacks-smart-dropout")
-
-
-def test_check_vttclean():
-    """There must not be VTT Talk sources in the font."""
-    check = CheckTester("vttclean")
-
-    good_font = TEST_FILE("mada/Mada-Regular.ttf")
-    assert_PASS(check(good_font))
-
-    bad_font = TEST_FILE("hinting/Roboto-VF.ttf")
-    assert_results_contain(check(bad_font), FAIL, "has-vtt-sources")
-
-
 def test_check_fvar_instances__another_test():  # TODO: REVIEW THIS.
     """Check variable font instances."""
     check = CheckTester("googlefonts:fvar_instances")
@@ -2869,24 +2825,6 @@ def NOT_IMPLEMENTED_test_check_family_tnum_horizontal_metrics():
     # code-paths:
     # - FAIL, "inconsistent-widths"
     # - PASS
-
-
-def test_check_integer_ppem_if_hinted():
-    """PPEM must be an integer on hinted fonts."""
-    check = CheckTester("integer_ppem_if_hinted")
-
-    # Our reference Merriweather Regular is hinted, but does not set
-    # the "rounded PPEM" flag (bit 3 on the head table flags) as
-    # described at https://docs.microsoft.com/en-us/typography/opentype/spec/head
-    ttFont = TTFont(TEST_FILE("merriweather/Merriweather-Regular.ttf"))
-
-    # So it must FAIL the check:
-    assert_results_contain(check(ttFont), FAIL, "bad-flags", "with a bad font...")
-
-    # hotfixing it should make it PASS:
-    ttFont["head"].flags |= 1 << 3
-
-    assert_PASS(check(ttFont), "with a good font...")
 
 
 def test_check_ligature_carets():
